## Deployment
This website is hosted in [Swiggy - Omkar](https://gregarious-crepe-46346c.netlify.app/). Just check it out.**Maybe it shows a warning because I didn't purchase an SSL certificate for this. However, you can still enter the website without any issue by clicking details**

## About
This repository contains a clone of the Swiggy front-end using HTML and CSS. The purpose of this project is to showcase my skills in web development and demonstrate my ability to recreate a popular website's user interface.

## Technologies Used
- HTML
- CSS

## Contributions
Contributions to this project are welcome. If you have any suggestions or improvements, feel free to create an issue or submit a pull request.

*Thanks a lot*